fix(#2806): R9b interlocking fail-open paths — receipt validation, dirty-SA exit, summary gate, anchored pass
