+++ crates/cli-sub-agent/src/pipeline_post_exec_dirty_sa.rs
@@ -2,6 +2,8 @@ use csa_session::{MetaSessionState, SessionResult};
 
 use super::PostExecContext;
 
+/// True for a SA-mode `run` that mutated files but produced no positive
+/// structured completion receipt. Such a run cannot be trusted as completed.
 pub(super) fn dirty_sa_run_lacks_completion_receipt(
     sa_mode: bool,
     task_type: Option<&str>,
@@ -14,6 +16,16 @@ pub(super) fn dirty_sa_run_lacks_completion_receipt(
         && !has_positive_structured_completion
 }
 
+/// Dirty-SA exit-preservation (#2806 R9b-F2): a SA-mode `run` that mutated
+/// files must NOT have a nonzero exit incidentally downgraded to success,
+/// regardless of whether a structured completion receipt is present. A valid
+/// receipt suppresses the exit-0 dirty-run-unconfirmed rewrite (handled by
+/// [`dirty_sa_run_lacks_completion_receipt`]), but it does not license
+/// downgrading a real nonzero failure on a dirty SA run.
+///
+/// This MUST run BEFORE the effective-outcome classification so it sees the raw
+/// nonzero exit. It records a CSA-own gate failure, which the classifier treats
+/// as authoritative-fatal (never downgraded).
 pub(super) fn maybe_mark_dirty_sa_run_without_receipt(
     ctx: &PostExecContext<'_>,
     session: &mut MetaSessionState,
@@ -21,6 +33,39 @@ pub(super) fn maybe_mark_dirty_sa_run_without_receipt(
     session_result: &mut SessionResult,
     has_positive_structured_completion: bool,
 ) {
+    // Dirty-SA exit-preservation: a nonzero exit on a dirty SA-mode run is
+    // authoritative. A valid receipt does not downgrade it (#2806 R9b-F2).
+    let dirty_sa_run =
+        ctx.sa_mode && matches!(ctx.task_type, None | Some("run")) && !ctx.changed_paths.is_empty();
+    if dirty_sa_run && result.exit_code != 0 {
+        // Capture the raw exit BEFORE marking the gate: `mark_gate_failure`
+        // forces exit_code to 1, but a timeout (124) / signal (137/143) exit
+        // must be preserved verbatim for status mapping and diagnostics
+        // (#2806 R9b-F2). The CSA gate marker makes the outcome classifier
+        // treat this as authoritative-fatal (never downgraded) without
+        // clobbering the specific terminal code. `session_result` was already
+        // initialized with the terminal-reason-derived status
+        // (initial_session_status, e.g. "timed_out"/"interrupted"), so preserve
+        // that status instead of recomputing from status_from_exit_code (which
+        // would map 124 -> "failure", losing the timeout/signal distinction).
+        let preserved_exit_code = result.exit_code;
+        tracing::warn!(
+            session = %session.meta_session_id,
+            exit_code = preserved_exit_code,
+            has_receipt = has_positive_structured_completion,
+            "Dirty SA-mode run exited nonzero — preserving nonzero exit (rejecting incidental downgrade)"
+        );
+        result.mark_gate_failure("dirty-sa-nonzero-exit-preserved");
+        result.exit_code = preserved_exit_code;
+        session_result.exit_code = preserved_exit_code;
+        if let Some(tool_state) = session.tools.get_mut(ctx.executor.tool_name()) {
+            tool_state.last_exit_code = preserved_exit_code;
+        }
+        return;
+    }
+
+    // Exit-0 dirty-run-unconfirmed: a SA-mode run that mutated files but lacks
+    // a positive structured completion receipt cannot be trusted as completed.
     if result.exit_code != 0
         || !dirty_sa_run_lacks_completion_receipt(
             ctx.sa_mode,
